+++ components/archive/search/src/main/java/com/flipperdevices/archive/search/viewmodel/SearchViewModel.kt
@@ -8,12 +8,14 @@ import com.flipperdevices.bridge.synchronization.api.SynchronizationApi
 import com.flipperdevices.bridge.synchronization.api.SynchronizationState
 import com.flipperdevices.keyparser.api.KeyParser
 import kotlinx.collections.immutable.toImmutableList
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.plus
 import tangle.viewmodel.VMInject
 
 class SearchViewModel @VMInject constructor(
@@ -31,7 +33,7 @@ class SearchViewModel @VMInject constructor(
                 .collect {
                     searchState.emit(SearchState.Loaded(it.toImmutableList()))
                 }
-        }.launchIn(viewModelScope)
+        }.launchIn(viewModelScope + Dispatchers.Default)
     }
 
     fun getState(): StateFlow<SearchState> = searchState
@@ -40,7 +42,7 @@ class SearchViewModel @VMInject constructor(
         synchronizationApi.getSynchronizationState()
 
     fun onChangeText(text: String) {
-        viewModelScope.launch {
+        viewModelScope.launch(Dispatchers.Default) {
             queryFlow.emit(text)
         }
     }