+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/viewmodel/BottomNavigationViewModel.kt
@@ -11,6 +11,7 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
 import com.flipperdevices.core.preference.pb.SelectedTab
 import com.flipperdevices.core.preference.pb.Settings
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.first
@@ -49,7 +50,7 @@ class BottomNavigationViewModel @VMInject constructor(
             return
         }
 
-        viewModelScope.launch {
+        viewModelScope.launch(Dispatchers.Default) {
             selectedTabInternal.emit(currentTab)
             settingsDataStore.updateData {
                 it.toBuilder()