+++ components/bridge/service/impl/src/main/java/com/flipperdevices/bridge/service/impl/delegate/FlipperActionNotifierImpl.kt
@@ -2,6 +2,7 @@ package com.flipperdevices.bridge.service.impl.delegate
 
 import com.flipperdevices.bridge.api.di.FlipperBleServiceGraph
 import com.flipperdevices.bridge.api.manager.delegates.FlipperActionNotifier
+import com.flipperdevices.core.di.SingleIn
 import com.squareup.anvil.annotations.ContributesBinding
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -10,6 +11,7 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
+@SingleIn(FlipperBleServiceGraph::class)
 @ContributesBinding(FlipperBleServiceGraph::class, FlipperActionNotifier::class)
 class FlipperActionNotifierImpl @Inject constructor(
     private val scope: CoroutineScope