+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [FIX] Endpoint for fap hub build download
 - [FIX] App open metric
 - [FIX] Remove kotlin incremental
+- [FIX] Faphub search, build status dialogs, placeholders
 - [Feature] Open application and screen streaming if install
 - [Feature] FapHub MVPv0
 - [Feature] Update README with Fdroid badge